backend/pyproject.toml
[project]
name = "quizik-api"
version = "0.1.0"
description = "Quizik — Quiz platform API"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Validation & settings
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    # Background jobs
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    # Logging
    "structlog>=24.0.0",
    # CORS / utilities
    "python-multipart>=0.0.9",
    # Parsers
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
    # Fuzzy matching for open-answer evaluation
    "rapidfuzz>=3.0.0",
    # Google OAuth token verification
    "google-auth>=2.0.0",
    "requests>=2.31.0",
    # Async HTTP (used for static file uploads and healthchecks)
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src", "services"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
